Extract logic into shared code
